Centralize policy modification
